--- app/src/interfaces/list-m2m-new/index.ts
@@ -1,90 +1,0 @@
-import { defineInterface } from '@directus/shared/utils';
-import InterfaceListM2M from './list-m2m.vue';
-import PreviewSVG from './preview.svg?raw';
-
-export default defineInterface({
-	id: 'list-m2m2',
-	name: '$t:interfaces.list-m2m.many-to-many2',
-	description: '$t:interfaces.list-m2m.description',
-	icon: 'note_add',
-	component: InterfaceListM2M,
-	relational: true,
-	types: ['alias'],
-	localTypes: ['m2m'],
-	group: 'relational',
-	options: ({ editing, relations }) => {
-		const displayTemplateMeta =
-			editing === '+'
-				? {
-						interface: 'presentation-notice',
-						options: {
-							text: '$t:interfaces.list-m2m.display_template_configure_notice',
-						},
-				  }
-				: {
-						interface: 'system-display-template',
-						options: {
-							collectionName: relations.o2m?.collection,
-						},
-				  };
-
-		return [
-			{
-				field: 'template',
-				name: '$t:display_template',
-				meta: displayTemplateMeta,
-			},
-			{
-				field: 'enableCreate',
-				name: '$t:creating_items',
-				schema: {
-					default_value: true,
-				},
-				meta: {
-					interface: 'boolean',
-					options: {
-						label: '$t:enable_create_button',
-					},
-					width: 'half',
-				},
-			},
-			{
-				field: 'enableSelect',
-				name: '$t:selecting_items',
-				schema: {
-					default_value: true,
-				},
-				meta: {
-					interface: 'boolean',
-					options: {
-						label: '$t:enable_select_button',
-					},
-					width: 'half',
-				},
-			},
-			{
-				field: 'filter',
-				name: '$t:filter',
-				type: 'json',
-				meta: {
-					interface: 'system-filter',
-					options: {
-						collectionName: relations.m2o?.related_collection ?? null,
-					},
-					conditions: [
-						{
-							rule: {
-								enableSelect: {
-									_eq: false,
-								},
-							},
-							hidden: true,
-						},
-					],
-				},
-			},
-		];
-	},
-	recommendedDisplays: ['related-values'],
-	preview: PreviewSVG,
-});
